# docs/swagger.yaml
basePath: /
definitions:
  dto.AddonItemResponse:
    properties:
      addonId:
        type: string
      quantity:
        type: integer
    type: object
  dto.AddonRequest:
    properties:
      addonId:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - addonId
    - quantity
    type: object
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.Metadata'
      success:
        type: boolean
    type: object
  dto.AvailabilityResponse:
    properties:
      date:
        description: YYYY-MM-DD
        type: string
      slots:
        items:
          $ref: '#/definitions/dto.AvailableSlotInfo'
        type: array
    type: object
  dto.AvailableSlotInfo:
    properties:
      available:
        description: true if slot is available
        type: boolean
      themeId:
        description: Theme ID for this slot (omitted when checking all themes)
        type: string
      time:
        description: HH:mm format
        type: string
    type: object
  dto.BookingRequest:
    properties:
      addons:
        items:
          $ref: '#/definitions/dto.AddonRequest'
        type: array
      customer:
        $ref: '#/definitions/dto.CustomerInfo'
      packageId:
        type: string
      slots:
        items:
          $ref: '#/definitions/dto.SlotRequest'
        maxItems: 3
        minItems: 1
        type: array
    required:
    - customer
    - packageId
    - slots
    type: object
  dto.BookingResponse:
    properties:
      addons:
        items:
          $ref: '#/definitions/dto.AddonItemResponse'
        type: array
      createdAt:
        description: ISO 8601 datetime
        type: string
      customer:
        $ref: '#/definitions/dto.CustomerInfo'
      id:
        type: string
      packageId:
        type: string
      paymentScreenshotUrl:
        type: string
      slots:
        items:
          $ref: '#/definitions/dto.SlotResponse'
        type: array
      status:
        description: PENDING | APPROVED | REJECTED
        type: string
      totalPrice:
        type: number
      updatedAt:
        description: ISO 8601 datetime
        type: string
    type: object
...
  dto.CustomerInfo:
    properties:
      email:
        type: string
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
  dto.Metadata:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
  dto.SlotRequest:
    properties:
      date:
        description: ISO date string (YYYY-MM-DD)
        type: string
      themeId:
        type: string
      time:
        description: Time in HH:mm format (e.g., "10:00")
        type: string
    required:
    - date
    - themeId
    - time
    type: object
  dto.SlotResponse:
    properties:
      date:
        description: ISO date string (YYYY-MM-DD)
        type: string
      themeId:
        description: Foreign key to Theme
        type: string
      time:
        description: Time in HH:mm format
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/handlers.UserInfo'
    type: object
  handlers.UserInfo:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@lqstudio.com
    name: API Support
  description: API for photography studio booking system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: LQ Studio Photography Booking API
  version: "1.0"
paths:
...
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin user and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Admin login
      tags:
      - auth
...
  /api/bookings:
    post:
      consumes:
      - application/json
      description: Create a new booking (customer-facing, no auth required)
      parameters:
      - description: Booking data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookingResponse'
              type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "409":
          description: Slot not available or conflict
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Create new booking
      tags:
      - bookings
  /api/bookings/{id}:
    get:
      consumes:
      - application/json
      description: Get a booking by ID (public - for customers to check their booking)
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookingResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get booking by ID
      tags:
      - bookings
  /api/bookings/{id}/payment-screenshot:
    post:
      consumes:
      - multipart/form-data
      description: Upload payment screenshot for a booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment screenshot image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BookingResponse'
              type: object
        "400":
          description: Invalid file or booking ID
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Booking not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "415":
          description: Invalid file type
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Upload payment screenshot
      tags:
      - bookings
...
  /api/themes/{themeId}/available-times:
    get:
      consumes:
      - application/json
      description: Check available time slots for a specific theme or all themes on
        a specific date. Use themeId="all" to check availability across all active
        themes.
      parameters:
      - description: Theme ID or 'all' for all themes
        in: path
        name: themeId
        required: true
        type: string
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        required: true
        type: string
      - description: Package ID (optional)
        in: query
        name: packageId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns availability with themeId in slots for specific theme,
            or without themeId for all themes
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AvailabilityResponse'
              type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Theme not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get available time slots
      tags:
      - availability
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"